+++ src/views/admin/default/MainPage.tsx
@@ -1,3 +1,4 @@
+/* eslint-disable @typescript-eslint/strict-boolean-expressions */
 // import MiniCalendar from 'components/calendar/MiniCalendar';
 // import WeeklyRevenue from 'views/admin/default/components/WeeklyRevenue';
 // import TotalSpent from 'views/admin/default/components/TotalSpent';
@@ -12,6 +13,7 @@ import ProjectCard from './components/ProjectCard';
 import LoadingComponent from './components/LoadingComponent ';
 import api from 'context/api';
 import { useQuery } from 'react-query';
+import axios from 'axios';
 
 enum UserRole {
    Subscriber = 'Subscriber',
@@ -192,6 +194,54 @@ const MainPage = (): JSX.Element => {
       navigate('/admin/project/create');
    }
 
+     /* department 설정 */
+     const [showDepartmentRegisterModal, setShowDepartmentRegisterModal] = React.useState(false);
+     const [isDepartment, setIsDepartment] = React.useState(false);
+     const [department, setDepartment] = React.useState(localStorage.getItem('department'));
+  
+     React.useEffect(() => {
+        if (localStorage.getItem('department') === 'null') {
+           console.log('showdepartment');
+           setShowDepartmentRegisterModal(true);
+        }
+     }, []);
+  
+     const handleModalClose = (): void => {
+        console.log(isDepartment);
+        setIsDepartment(false);
+     };
+  
+     const handleSelectUserDepartmentChange = (event: React.ChangeEvent<HTMLSelectElement>): void => {
+        const { value } = event.target;
+        setDepartment(value);
+     };
+
+     const handleClickSaveChangeButton = (): void => {
+      axios
+         .patch(
+            'http://localhost:8080/api/user',
+            {
+               department: department,
+            },
+            {
+               headers: {
+                  Authorization: localStorage.getItem('token'),
+               },
+            },
+         )
+         .then(response => {
+            console.log(response.data); // Process the response as needed
+            localStorage.setItem('department', department ?? '');
+            setDepartment(localStorage.getItem('department'));
+         })
+         .catch(error => {
+            console.error(error);
+            // Handle error cases here
+         });
+      setShowDepartmentRegisterModal(false);
+   };
+
+
    return (
       <div>
          <div className="pt-3">
@@ -335,7 +385,27 @@ const MainPage = (): JSX.Element => {
             <RecentRelease />
          </div>
 
-        
+        {showDepartmentRegisterModal && (
+            <div className="userprofileCard-modal-overlay">
+               <div className="userprofileCard-modal-content">
+                  <span className="userprofileCard-close" onClick={handleModalClose}>
+                     &times;
+                  </span>
+                  <h2>Select Department</h2>
+                  {/* Department selection options */}
+                  <div className="modal-body">
+                     <select name="department" value={department ?? ''} onChange={handleSelectUserDepartmentChange}>
+                        <option value="Department 1">Department 1</option>
+                        <option value="Department 2">Department 2</option>
+                        <option value="Department 3">Department 3</option>
+                     </select>
+                  </div>
+                  <button type="button" onClick={handleClickSaveChangeButton}>
+                     Save
+                  </button>
+               </div>
+            </div>
+         )}
       </div>
    );
 };
